Mayor Feature: * Code modularization by functions. * Fix VEN format currency.
