+++ app/functions/config.py
@@ -1,0 +1,13 @@
+country_code_dict = {
+    'ARG': 'ARG',
+    'COP': 'COP',
+    'BRL': 'BRL',
+    'VEN': 'VEN',
+    'DOP_POPU': 'DOP_POPU',
+    'DOP_BANRE': 'DOP_BANRE',
+}
+
+return_methods = {
+    'json': 'json',
+    'csv': 'csv',
+}
